Smart rate aded + Request manage sort

+++ DataLayer/Entities/Service.cs
@@ -13,7 +13,7 @@ namespace DataLayer.Entities
         public string Name { get; set; }
         public string Description { get; set; }
         public int EventsPerDay { get; set; }
-        public float SmartRate { get; set; }
+        public double SmartRate { get; set; }
         public int NumberOfRatings { get; set; }
         public bool IsDeleted { get; set; }
     }
